# [WIP] port selected Kokkos Core performance test