--- a/algorithms/src/Kokkos_NestedSort.hpp
+++ b/algorithms/src/Kokkos_NestedSort.hpp
@@ -87,11 +87,10 @@ struct NestedRange<false> {
 // This only takes the NestedRange instance for template arg deduction.
 template <class TeamMember, class KeyViewType, class ValueViewType,
           class Comparator, bool useTeamLevel>
-KOKKOS_INLINE_FUNCTION void sort_nested_impl(const TeamMember& t,
-                                             const KeyViewType& keyView,
-                                             const ValueViewType& valueView,
-                                             const Comparator& comp,
-                                             const NestedRange<useTeamLevel>) {
+KOKKOS_INLINE_FUNCTION void sort_nested_impl(
+    const TeamMember& t, const KeyViewType& keyView,
+    [[maybe_unused]] const ValueViewType& valueView, const Comparator& comp,
+    const NestedRange<useTeamLevel>) {
   using SizeType  = typename KeyViewType::size_type;
   using KeyType   = typename KeyViewType::non_const_value_type;
   using Range     = NestedRange<useTeamLevel>;