--- a/algorithms/src/Kokkos_Sort.hpp
+++ b/algorithms/src/Kokkos_Sort.hpp
@@ -50,8 +50,8 @@
 #endif
 
 #include <Kokkos_Core.hpp>
-#include <algorithm>
 #include <Kokkos_NestedSort.hpp>
+#include <algorithm>
 
 namespace Kokkos {