--- a/algorithms/unit_tests/TestNestedSort.hpp
+++ b/algorithms/unit_tests/TestNestedSort.hpp
@@ -47,6 +47,8 @@
 
 #include <unordered_set>
 #include <random>
+#include <Kokkos_Random.hpp>
+#include <Kokkos_NestedSort.hpp>
 
 namespace Test {