--- a/core/perf_test/PerfTest_ViewCopy.cpp
+++ b/core/perf_test/PerfTest_ViewCopy.cpp
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include <Benchmark_ViewCopy.hpp>
+#include <PerfTest_ViewCopy.hpp>
 
 namespace Test {