… .github/workflows/continuous-integration-workflow.yml
… CMakeLists.txt
… algorithms/unit_tests/CMakeLists.txt
… algorithms/unit_tests/TestSort.hpp
… algorithms/unit_tests/TestSortCommon.hpp
… appveyor.yml
… bin/hpcbind
… cmake/KokkosCore_config.h.in
… cmake/kokkos_arch.cmake
… cmake/kokkos_compiler_id.cmake
… cmake/kokkos_enable_devices.cmake
… cmake/kokkos_enable_options.cmake
… cmake/kokkos_functions.cmake
… cmake/kokkos_tribits.cmake
… containers/performance_tests/TestHIP.cpp
… containers/src/Kokkos_DualView.hpp
… containers/src/Kokkos_DynRankView.hpp
… containers/src/Kokkos_DynamicView.hpp
… containers/src/Kokkos_OffsetView.hpp
… containers/src/Kokkos_ScatterView.hpp
… containers/unit_tests/TestDualView.hpp
… containers/unit_tests/TestErrorReporter.hpp
… containers/unit_tests/TestScatterView.hpp
… containers/unit_tests/TestWithoutInitializing.hpp
… core/perf_test/BenchmarkMain.cpp
… core/perf_test/Benchmark_Context.hpp
… core/perf_test/Benchmark_Sample.cpp
… core/perf_test/Benchmark_ViewCopy.cpp
… core/perf_test/Benchmark_ViewCopy.hpp
… core/perf_test/CMakeLists.txt