Local draft PR to avoid polluting Kokkos main repo.
Note: targets [integrate-google-benchmark](https://github.com/cz4rs/kokkos/tree/integrate-google-benchmark) branch to facilitate discussion.

---
Current sample output:
```
kokkos  OMP_PROC_BIND=spread OMP_PLACES=threads ./build/core/perf_test/KokkosCore_PerformanceTest_Benchmark
```
```
2022-06-08T17:44:54+02:00
Running ./build/core/perf_test/KokkosCore_PerformanceTest_Benchmark
Run on (8 X 4700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 12288 KiB (x1)
Load Average: 2.13, 1.42, 1.26
Kokkos configuration: Kokkos Version: 3.6.99
Compiler:
KOKKOS_COMPILER_CLANG: 1400
Architecture:
Default Device: N6Kokkos6OpenMPE
Atomics:
KOKKOS_ENABLE_GNU_ATOMICS: no
KOKKOS_ENABLE_INTEL_ATOMICS: no
KOKKOS_ENABLE_WINDOWS_ATOMICS: no
Vectorization:
KOKKOS_ENABLE_PRAGMA_IVDEP: no
KOKKOS_ENABLE_PRAGMA_LOOPCOUNT: no
KOKKOS_ENABLE_PRAGMA_SIMD: no
KOKKOS_ENABLE_PRAGMA_UNROLL: no
KOKKOS_ENABLE_PRAGMA_VECTOR: no
Memory:
KOKKOS_ENABLE_HBWSPACE: no
KOKKOS_ENABLE_INTEL_MM_ALLOC: no
Options:
KOKKOS_ENABLE_ASM: no
KOKKOS_ENABLE_CXX14: yes
KOKKOS_ENABLE_CXX17: no
KOKKOS_ENABLE_CXX20: no
KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK: yes
KOKKOS_ENABLE_HWLOC: yes
KOKKOS_ENABLE_LIBDL: yes
KOKKOS_ENABLE_LIBRT: no
Host Parallel Execution Space:
  KOKKOS_ENABLE_OPENMP: yes
OpenMP Atomics:
  KOKKOS_ENABLE_OPENMP_ATOMICS: no

OpenMP Runtime Configuration:
Kokkos::OpenMP thread_pool_topology[ 1 x 8 x 1 ]


***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                     Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------------------------------------------------------
ViewDeepCopy_Rank8<Kokkos::LayoutRight, Kokkos::LayoutLeft>/N:10/R:1/manual_time           2.96 s          2.95 s             1 GB/s=0.502777 MB=762.939
ViewDeepCopy_Rank8_Raw<Kokkos::LayoutRight, Kokkos::LayoutLeft>/N:10/R:1/manual_time      0.083 s         0.083 s             9 GB/s=18.5656 MB=762.939
```

<details>
  <summary>JSON output</summary>
  
```
{
  "context": {
    "date": "2022-06-10T17:39:29+02:00",
    "host_name": "rysy",
    "executable": "./build/core/perf_test/KokkosCore_PerformanceTest_Benchmark",
    "num_cpus": 8,
    "mhz_per_cpu": 4700,
    "cpu_scaling_enabled": true,
    "caches": [
      {
        "type": "Data",
        "level": 1,
        "size": 32768,
        "num_sharing": 1
      },
      {
        "type": "Instruction",
        "level": 1,
        "size": 32768,
        "num_sharing": 1
      },
      {
        "type": "Unified",
        "level": 2,
        "size": 262144,
        "num_sharing": 1
      },
      {
        "type": "Unified",
        "level": 3,
        "size": 12582912,
        "num_sharing": 8
      }
    ],
    "load_avg": [2.54102,1.82324,1.64014],
    "library_build_type": "debug",
    "Kokkos configuration": "Kokkos Version: 3.6.99\nCompiler:\nKOKKOS_COMPILER_CLANG: 1400\nArchitecture:\nDefault Device: N6Kokkos6OpenMPE\nAtomics:\nKOKKOS_ENABLE_GNU_ATOMICS: no\nKOKKOS_ENABLE_INTEL_ATOMICS: no\nKOKKOS_ENABLE_WINDOWS_ATOMICS: no\nVectorization:\nKOKKOS_ENABLE_PRAGMA_IVDEP: no\nKOKKOS_ENABLE_PRAGMA_LOOPCOUNT: no\nKOKKOS_ENABLE_PRAGMA_SIMD: no\nKOKKOS_ENABLE_PRAGMA_UNROLL: no\nKOKKOS_ENABLE_PRAGMA_VECTOR: no\nMemory:\nKOKKOS_ENABLE_HBWSPACE: no\nKOKKOS_ENABLE_INTEL_MM_ALLOC: no\nOptions:\nKOKKOS_ENABLE_ASM: no\nKOKKOS_ENABLE_CXX14: yes\nKOKKOS_ENABLE_CXX17: no\nKOKKOS_ENABLE_CXX20: no\nKOKKOS_ENABLE_DEBUG_BOUNDS_CHECK: yes\nKOKKOS_ENABLE_HWLOC: yes\nKOKKOS_ENABLE_LIBDL: yes\nKOKKOS_ENABLE_LIBRT: no\nHost Parallel Execution Space:\n  KOKKOS_ENABLE_OPENMP: yes\nOpenMP Atomics:\n  KOKKOS_ENABLE_OPENMP_ATOMICS: no\n\nOpenMP Runtime Configuration:\nKokkos::OpenMP thread_pool_topology[ 1 x 8 x 1 ]\n\n"
  },
  "benchmarks": [
    {
      "name": "ViewDeepCopy_Rank8<Kokkos::LayoutRight, Kokkos::LayoutLeft>/N:10/R:1/manual_time",
      "family_index": 0,
      "per_family_instance_index": 0,
      "run_name": "ViewDeepCopy_Rank8<Kokkos::LayoutRight, Kokkos::LayoutLeft>/N:10/R:1/manual_time",
      "run_type": "iteration",
      "repetitions": 1,
      "repetition_index": 0,
      "threads": 1,
      "iterations": 1,
      "real_time": 3.3720064879999998e+00,
      "cpu_time": 3.1431697110000001e+00,
      "time_unit": "s",
      "GB/s": 4.4190784468762440e-01,
      "MB": 7.6293945312500000e+02
    },
    {
      "name": "ViewDeepCopy_Rank8_Raw<Kokkos::LayoutRight, Kokkos::LayoutLeft>/N:10/R:1/manual_time",
      "family_index": 1,
      "per_family_instance_index": 0,
      "run_name": "ViewDeepCopy_Rank8_Raw<Kokkos::LayoutRight, Kokkos::LayoutLeft>/N:10/R:1/manual_time",
      "run_type": "iteration",
      "repetitions": 1,
      "repetition_index": 0,
      "threads": 1,
      "iterations": 8,
      "real_time": 8.6857682374999995e-02,
      "cpu_time": 8.6535297374999987e-02,
      "time_unit": "s",
      "GB/s": 1.7332003851833822e+01,
      "MB": 7.6293945312500000e+02
    }
  ]
}
```
</details>



Original test:
```
kokkos  OMP_PROC_BIND=spread OMP_PLACES=threads ./build/core/perf_test/KokkosCore_PerfTestExec --gtest_filter=*ViewDeepCopy_RightLeft_Rank8*
```
```
Note: Google Test filter = *ViewDeepCopy_RightLeft_Rank8*
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from default_exec
[ RUN      ] default_exec.ViewDeepCopy_RightLeft_Rank8
DeepCopy Performance for LayoutRight to LayoutLeft:
   Raw:   0.080353 s   762.939453 MB   18.544586 GB/s
   Rank8: 3.017889 s   762.939453 MB   0.493761 GB/s
[       OK ] default_exec.ViewDeepCopy_RightLeft_Rank8 (3979 ms)
[----------] 1 test from default_exec (3979 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test suite ran. (3979 ms total)
[  PASSED  ] 1 test.
```